forced real to integer because it is being used as an index. pathscale compiler on lightning complained ... needed nint(highest_obs_level). (CVS 1.57)
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@2551 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 34aa637d1ff9ea7aea286880de0adbdd4789e357

+++ models/cam/model_mod.f90
@@ -2857,7 +2857,7 @@ else
       call plevs_cam (p_surf, num_levs, pfull)
       highest_obs_level = 1
       threshold = highest_obs_pressure_mb*100.0_r8
-      do while ((pfull(highest_obs_level)) < threshold )
+      do while ((pfull(nint(highest_obs_level))) < threshold )
          highest_obs_level = highest_obs_level +1
       end do
       deallocate(pfull)
